+++ Trainer_v4/Main.cs
@@ -8,7 +8,7 @@ namespace Trainer_v4
     {
         private TrainerBehaviour _trainerBehaviour;
         public static Button btn;
-        public static string version = "(v4.2)";
+        public static string version = "(v4.3)";
 
         public static bool IsShowed
         {